Add the blakey generator based on BLAKE3
