+++ tools/archiver.py
@@ -499,7 +499,7 @@ class Archiver(object):  # N.B. Also used by import-mbox.py
 
         return output_json, contents, msg_metadata, irt
 
-    def archive_message(self, args, mlist, msg, raw_message):
+    def archive_message(self, mlist, msg, raw_message=None, dry=False, dump=False):
         """Send the message to the archiver.
 
         :param args: Command line args (verbose, ibody)
@@ -523,6 +523,8 @@ class Archiver(object):  # N.B. Also used by import-mbox.py
         ):
             private = True
 
+        if raw_message is None:
+            raw_message = msg.as_bytes()
         ojson, contents, msg_metadata, irt = self.compute_updates(
             lid, private, msg, raw_message
         )
@@ -530,18 +532,18 @@ class Archiver(object):  # N.B. Also used by import-mbox.py
             _id = msg.get("message-id") or msg.get("Subject") or msg.get("Date")
             raise Exception("Could not parse message %s for %s" % (_id, lid))
 
-        if args.dry:
+        if dry:
             print("**** Dry run, not saving message to database *****")
             return lid, ojson["mid"]
 
-        if args.dump:
+        if dump:
             try:
                 elastic = plugins.elastic.Elastic()
             except elasticsearch.exceptions.ElasticsearchException as e:
                 print(e)
                 print(
                     "ES connection failed, but dumponfail specified, dumping to %s"
-                    % args.dump
+                    % dump
                 )
         else:
             elastic = plugins.elastic.Elastic()
@@ -588,12 +590,12 @@ class Archiver(object):  # N.B. Also used by import-mbox.py
         # We'll leave it to another process to pick up the slack.
         except Exception as err:
             print(err)
-            if args.dump:
+            if dump:
                 print(
                     "Pushing to ES failed, but dumponfail specified, dumping JSON docs"
                 )
                 uid = uuid.uuid4()
-                mbox_path = os.path.join(args.dump, "%s.json" % uid)
+                mbox_path = os.path.join(dump, "%s.json" % uid)
                 with open(mbox_path, "w") as f:
                     json.dump(
                         {
@@ -898,7 +900,7 @@ def main():
             )
 
             try:
-                lid, mid = archie.archive_message(args, list_data, msg, raw_message)
+                lid, mid = archie.archive_message(list_data, msg, raw_message, dry=args.dry, dump=args.dump)
                 print(
                     "%s: Done archiving to %s as %s!"
                     % (email.utils.formatdate(), lid, mid)
